update: use dynamic import on ProductItem, ProductWishlistButton, and AuthenitcationButtons
